Moved things to pages directory and finished member display
