+++ client/src/pages/landingpage/DevPartners.js
