+++ client/src/pages/landingpage/citation/Partner.js
