+++ client/src/pages/landingpage/citation/PartnerGeneration.js
