+++ client/src/pages/landingpage/citation/SlideSet.js
