+++ client/src/pages/landingpage/error.js
@@ -1,7 +1,7 @@
 /* eslint-disable new-cap */
 /* eslint-disable react/prop-types */
 import React from "react";
-import Redirect from "../../../helpers/Redirect";
+import Redirect from "../../helpers/Redirect";
 
 const Error = (props) => {
   const { actions } = props;
